refactor: Move ConsumedProduct entity to /TrackerAndMission (PLTR-90)
