+++ server/plantry/src/main/java/com/gdscplantry/plantry/domain/Pantry/service/ProductService.java
@@ -3,10 +3,14 @@ package com.gdscplantry.plantry.domain.Pantry.service;
 import com.gdscplantry.plantry.domain.Notification.domain.Notification;
 import com.gdscplantry.plantry.domain.Notification.domain.NotificationRepository;
 import com.gdscplantry.plantry.domain.Notification.service.RelatedNotificationService;
-import com.gdscplantry.plantry.domain.Pantry.domain.*;
+import com.gdscplantry.plantry.domain.Pantry.domain.Product;
+import com.gdscplantry.plantry.domain.Pantry.domain.ProductRepository;
+import com.gdscplantry.plantry.domain.Pantry.domain.UserPantryRepository;
 import com.gdscplantry.plantry.domain.Pantry.dto.product.*;
 import com.gdscplantry.plantry.domain.Pantry.error.PantryErrorCode;
 import com.gdscplantry.plantry.domain.Pantry.vo.FoodDataVo;
+import com.gdscplantry.plantry.domain.TrackerAndMission.domain.tracker.ConsumedProduct;
+import com.gdscplantry.plantry.domain.TrackerAndMission.domain.tracker.ConsumedProductRepository;
 import com.gdscplantry.plantry.domain.User.domain.User;
 import com.gdscplantry.plantry.domain.model.ProductDeleteTypeEnum;
 import com.gdscplantry.plantry.domain.model.StorageEnum;
